PP-13733 Extract types to separate file
